+++ example-context/example-persistence/src/test/java/uk/gov/justice/services/example/cakeshop/persistence/CakeOrderRepositoryIT.java
@@ -24,7 +24,6 @@ public class CakeOrderRepositoryIT extends BaseTransactionalTest {
     @Inject
     private CakeOrderRepository cakeOrderRepository;
 
-    private CakeOrder cakeOrderA;
 
     @Test
     public void shouldStoreOrder() {