+++ example-context/example-service/example-event/example-event-indexer/src/main/java/uk/gov/justice/services/example/cakeshop/dummy/DummyUnifiedSearchIndexer.java
@@ -1,0 +1,29 @@
+package uk.gov.justice.services.example.cakeshop.dummy;
+
+import static java.util.UUID.fromString;
+
+import uk.gov.justice.services.example.cakeshop.persistence.IndexRepository;
+import uk.gov.justice.services.example.cakeshop.persistence.entity.Index;
+import uk.gov.justice.services.unifiedsearch.UnifiedSearchIndexer;
+
+import java.time.ZonedDateTime;
+import java.util.UUID;
+
+import javax.enterprise.context.ApplicationScoped;
+import javax.inject.Inject;
+import javax.json.JsonObject;
+
+@ApplicationScoped
+public class DummyUnifiedSearchIndexer implements UnifiedSearchIndexer {
+
+    @Inject
+    private IndexRepository indexRepository;
+
+    @Override
+    public void indexData(final JsonObject jsonObject) {
+        final UUID recipeId = fromString(jsonObject.getString("recipeId"));
+        final ZonedDateTime deliveryDate = ZonedDateTime.parse(jsonObject.getString("deliveryDate"));
+        final Index index = new Index(recipeId, deliveryDate);
+        indexRepository.save(index);
+    }
+}